pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Aleks-Ti"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "1.0.0"
bitarray = "2.8.0"
aiogram = "^3.4.1"
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pillow = "^10.3.0"
redis = "^5.0.7"
celery = "^5.4.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.5"
ipykernel = "^6.29.5"

[tool.ruff]
line-length = 135
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py312"
select = ["E", "W", "F", "B", "I", "Q", "COM"]
fixable = ["I", "W", "COM", "Q"]
flake8-quotes.inline-quotes = "double"
flake8-quotes.docstring-quotes = "double"
ignore = ["D100"]
exclude = [
    "settings.py",
    "manage.py",
    "migrations",
    "__init__.py",
    ".cache",
    "__pycache__",
    "infra",
    "nginx",
    ".vscode",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "env.py",
    "versions",
    "migrations",
]

[tool.ruff.per-file-ignores]
"src/settings.py" = ["E501"]

[tool.ruff.pydocstyle]
convention = "google"